feat: split code to manageable chunks
